+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - when verifier proposing, `Chain` checks, if verifier is active instead
   of if it is only created.
 - split Chain contract into two separate contracts (storage and manager)
+- Flexible duration of election phases
 
 ## [0.3.0] - 2018-12-13
 ### Added: